+++ CHANGELOG.md
@@ -8,6 +8,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 
+### Added
+- The CLI class `EditConfiguration`, which edits the configuration file.
+
+### Changed
+- Section for `ActionCli` decorators over-write that which is defined at the
+  class level for an application in `CLI_META`.
+
+
 ## [1.7.1] - 2022-01-12
 ### Changed
 - Shelve extension calculation is not heuristically calculated by creating a